fix(wrap) rename helper, move wrap insertion to helper
